requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
exceptiongroup==1.1.2
fastapi==0.101.0
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
lxml==4.9.3
multidict==6.0.4
openai==0.27.8
pydantic==2.1.1
pydantic_core==2.4.0
pypdf==3.15.0
python-docx==0.8.11
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
yarl==1.9.2
